requirements/base.txt
dj-database-url
Django
django-environ
djangorestframework
drf-nested-routers
psycopg2
psycopg2-binary
pytz
six
